pyproject.toml
[tool.poetry]
name = "linkfix-for-discord"
version = "2.0.0-alpha"
description = "LinkFix for Discord replies to messages containing URLS that don't embed properly (x.com, twitter.com, YouTube Shorts) with URLS that do (fxtwitter.com, youtu.be)."
authors = ["Ralph <ralph@podaboutli.st>"]
readme = "README.md"

[tool.poe.tasks]
lint-ruff = "ruff check --verbose src/"
lint-black = "black --check --verbose src/"
lint = ["lint-ruff", "lint-black"]

[tool.ruff]
select = ["B", "D", "E", "F", "I", "N", "S", "W", "UP"]
ignore = ["D211", "D213", "D400"]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.format]
line-ending = "lf"


[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "==1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.292"
pre-commit = "^3.4.0"

[tool.poetry.group.optional]
optional = true

[tool.poetry.group.optional.dependencies]
poethepoet = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
